third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/css-rule-hover-highlights-selectors-expected.txt


Running: setupProxyOverlay

Running: testRegularNodeSelection
Highlights drawn: 4

Running: testShadowDOMNodeSelection
Highlights drawn: 6
